+++ src/public-api.ts
@@ -4,11 +4,7 @@ export * from './core';
 export * from './forms-sql/FormQueryParser';
 export * from './forms-sql/SqlRenderer';
 
-export * from './express-router/ExpressRouter';
-export * from './express-router/ApiModule';
-
-export * from './events/PublicEventServer';
-
+export * from './express-router';
 export * from './tokens';
 export * from './servers';
 export * from './services';